requirements.txt
streamlit
wikipedia-api
requests
